llm-benchmark/requirements.txt
# LLM Benchmark System Requirements

# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
PyYAML==6.0.1

# LLM Client (install from local lib directory)
# Run: pip install -e lib/rotator_library
litellm>=1.30.0

# Terminal UI
rich==13.7.0

# Web Viewer
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.6

# Optional: for better performance
# orjson==3.9.15
